+++ src/main/java/lotto/view/message/OutputMessage.java
@@ -1,7 +1,6 @@
 package lotto.view.message;
 
 import static lotto.utils.constants.LottoConstants.MATCH_BONUS_MESSAGE_FORMAT;
-import static lotto.utils.constants.LottoConstants.MATCH_FIVE_AND_BONUS_COUNT;
 import static lotto.utils.constants.LottoConstants.MATCH_FIVE_AND_BONUS_PRICE;
 import static lotto.utils.constants.LottoConstants.MATCH_FIVE_COUNT;
 import static lotto.utils.constants.LottoConstants.MATCH_FIVE_PRICE;
@@ -37,7 +36,7 @@ public enum OutputMessage {
         THREE_MATCH(String.format(MATCH_MESSAGE_FORMAT, MATCH_THREE_COUNT, MATCH_THREE_PRICE)),
         FOUR_MATCH(String.format(MATCH_MESSAGE_FORMAT, MATCH_FOUR_COUNT, MATCH_FOUR_PRICE)),
         FIVE_MATCH(String.format(MATCH_MESSAGE_FORMAT, MATCH_FIVE_COUNT, MATCH_FIVE_PRICE)),
-        FIVE_AND_BONUS_MATCH(String.format(MATCH_BONUS_MESSAGE_FORMAT, MATCH_FIVE_AND_BONUS_COUNT, MATCH_FIVE_AND_BONUS_PRICE)),
+        FIVE_AND_BONUS_MATCH(String.format(MATCH_BONUS_MESSAGE_FORMAT, MATCH_FIVE_COUNT, MATCH_FIVE_AND_BONUS_PRICE)),
         SIX_MATCH(String.format(MATCH_MESSAGE_FORMAT, MATCH_SIX_COUNT, MATCH_SIX_PRICE));
 
         private final String detail;
